requirements.txt
django==4.1.2
django-tinymce
pyrebase
numpy==1.21.5
pandas~=1.4.2
yfinance
sklearn
requests~=2.27.1
beautifulsoup4~=4.11.1
gnews
newspaper3k
python-dateutil~=2.8.2
finviz
htmldate
newsapi